+++ src/reputation/models.py
@@ -1,9 +1,9 @@
-from reputation.related_models.author_rsc import Escrow
 from reputation.related_models.bounty import Bounty
 from reputation.related_models.contribution import Contribution
 from reputation.related_models.deposit import Deposit
 from reputation.related_models.distribution import Distribution
 from reputation.related_models.distribution_amount import DistributionAmount
+from reputation.related_models.escrow import Escrow
 from reputation.related_models.paid_status_mixin import PaidStatusModelMixin
 from reputation.related_models.term import Term
 from reputation.related_models.webhook import Webhook
